+++ src/main/java/com/example/demo/model/ActiveFisic.java
@@ -1,5 +1,7 @@
 package com.example.demo.model;
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.*;
 
 @Entity
@@ -19,7 +21,8 @@ public class ActiveFisic extends ActiveBase {
     @Column(length = 20)
     private EstadoActivo estado;
 
-    public ActiveFisic(long id, String name, String descripcion, Double valor, EstadoActivo estatus, String ubicacion) {
+    public ActiveFisic(long id, String name, String descripcion, BigDecimal valor, EstadoActivo estatus,
+            String ubicacion) {
         super(id, name, descripcion, valor);
         this.estado = estatus;
         this.ubicacion = ubicacion;
